Refactored the dev server into several components
