+++ lib/lingon.js
@@ -14,6 +14,8 @@ var help              = require('./utils/help');
 var buildTask         = require('./tasks/build');
 var cleanTask         = require('./tasks/clean');
 var serverTask        = require('./tasks/server');
+var staticServerTask  = require('./tasks/static');
+
 var Builder           = require('./builder');
 
 var Lingon = function (rootPath, argv) {
@@ -44,6 +46,7 @@ var Lingon = function (rootPath, argv) {
   // Set up default tasks
   buildTask(this);
   serverTask(this);
+  staticServerTask(this);
   cleanTask(this);
 
   // Set up default help
