+++ lib/server/dynamicRequestHandler.js
@@ -1,32 +1,11 @@
-'use strict';
-
+var url               = require('url');
 var path              = require('path');
 var fs                = require('fs');
-var url               = require('url');
+var log               = require('../utils/log');
+var ExtensionRewriter = require('../utils/extensionRewriter');
+var buildCallback     = require('./buildCallback');
 
-var express           = require('express');
-var send              = require('send');
-var chalk             = require('chalk');
-
-var log               = require('./utils/log');
-var ExtensionRewriter = require('./utils/extensionRewriter');
-
-var server = function (lingon, ip, port) {
-  var app = express();
-
-  lingon.server = app;
-
-  var buildCallback = function (request, response, next) {
-    return function (requestFiles) {
-      if (!requestFiles || !requestFiles[0]) { return next(); }
-
-      var file = requestFiles[0];
-      response.type(send.mime.lookup(file.path));
-      response.body = file.contents;
-
-      next();
-    };
-  };
+var dynamicRequestHandlerFactory = function (lingon, ip, port) {
 
   var pathCache = {};
 
@@ -119,49 +98,7 @@ var server = function (lingon, ip, port) {
     });
   };
 
-  var catchAllHandler = function (request, response, next) {
-    if (!response.body && lingon.config.server.catchAll) {
-      lingon.build({
-        callback: buildCallback(request, response, next),
-        requestPath: lingon.config.server.catchAll,
-        pipelineTerminators: [],
-      });
-    } else {
-      next();
-    }
-  };
-
-  var responseHandler = function (request, response, next) {
-    if (response.body) {
-      return response.send(response.body);
-    }
-
-    response.status(404).send('File not found');
-  };
-
-  app.use(lingon.config.server.namespace, requestHandler);
-  lingon.trigger('serverConfigure');
-  app.use(catchAllHandler);
-  app.use(responseHandler);
-
-  process.on('uncaughtException', function (error) {
-    if (error.code == 'EADDRINUSE') {
-      log.error('A port is already in use, lingon could not start properly!');
-      log.info('[Info] Try starting lingon with a different port (' +
-          chalk.blue('lingon server -p <PORT>') +
-          ') or check your plugin configurations.');
-    } else {
-      log.error(error.message);
-    }
-  });
-
-  app.listen(port, ip, function () {
-    var hostname = ip == '0.0.0.0' ? 'localhost' : ip; // chrome started disallowing 0.0.0.0 in the browser URL
-    log.info('http server listening on: http://' + hostname + ':' + port +
-        lingon.config.server.namespace);
-    lingon.trigger('serverStarted');
-  });
-
+  return requestHandler;
 };
 
-module.exports = server;
+module.exports = dynamicRequestHandlerFactory;
