+++ lib/tasks/static.js
@@ -3,13 +3,13 @@
 var server = require('../server');
 
 module.exports = function (lingon) {
-  lingon.registerTask('server', function (callback) {
+  lingon.registerTask('static', function (callback) {
     var serverIp = lingon.argv.bind || '0.0.0.0';
     var serverPort = lingon.argv.p || '5678';
-    server(lingon, serverIp, serverPort, false);
+    server(lingon, serverIp, serverPort, true);
   }, {
 
-    message: 'Start the HTTP server',
+    message: 'Build once and serve static files',
     arguments: {
       p: 'Run on custom port',
     },
